% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum.R
\name{sum}
\alias{sum}
\title{Calculates the sum of the given 2 variables}
\usage{
sum(a, b)
}
\arguments{
\item{a}{: numeric variable}

\item{b}{: numeric variable}
}
\description{
Calculates the sum of the given 2 variables
}
